Inline `aircraft-layer` code
