--- app/layers/index.js
@@ -1,1 +1,0 @@
-export { default as AircraftLayer } from './aircraft';
